requirements.txt
pandas
argparse
pyarrow
openai
python-dotenv